+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fixed errors in Inner Radiance Torrents
 - Ghostly Carrier
+- BG3 Style Force Barrage
 
 ## [0.58.0] - 2025-11-11